Adjust font as well as solarized theme to work for both GUI & console
I had to add the emacs-color-theme-solarized package to the repository
because it was the only stable way to get emacs to reliably load the
file.

This is because I experienced issues in failures to apply the theme
colors at times citing "Could not load color..." and at other times it
would fail because it would try to load `color-themes` for some reason,
which doesn't seem to be necessary when using Emacs 24. At least this
way I know it will work as advertised.